Moved the getTokenType tests from ui to commons
